requirements-test.txt
-c requirements.txt
pytest
pytest-django
pytest-freezegun
